Update PKGBUILD
Now I am not sure, first there is a conflict between the two, where /usr/lib/libreadline.so exists in both pkgs, the the libc.so is not found. it still builds but whether it is right I am not sure.


Great work man, why don't you push some of these to testing at least so it makes it easier to keep a clean chroot for each build?